move service cls and convert token to secret str
