+++ openhands/server/auth.py
@@ -1,9 +1,10 @@
 from fastapi import Request
+from pydantic import SecretStr
 
 
 def get_github_token(request: Request) -> str | None:
     return getattr(request.state, 'github_token', None)
 
 
-def get_user_id(request: Request) -> str | None:
+def get_user_id(request: Request) -> SecretStr | None:
     return getattr(request.state, 'github_user_id', None)
