+++ openhands/server/middleware.py
@@ -196,7 +196,7 @@ class GitHubTokenMiddleware(SessionMiddlewareInterface):
         # TODO: To avoid checks like this we should re-add the abilty to have completely different middleware in SAAS as in OSS
         if getattr(request.state, 'github_token', None) is None:
             if settings and settings.github_token:
-                request.state.github_token = settings.github_token.get_secret_value()
+                request.state.github_token = settings.github_token
             else:
                 request.state.github_token = None
