+++ openhands/server/routes/github.py
@@ -1,11 +1,12 @@
 from fastapi import APIRouter, Depends
 from fastapi.responses import JSONResponse
+from pydantic import SecretStr
 
-from openhands.server.auth import get_user_id
+from openhands.server.auth import get_github_token, get_user_id
 from openhands.server.data_models.gh_types import GitHubRepository, GitHubUser
-from openhands.server.services.github_service import GitHubService
 from openhands.server.shared import server_config
 from openhands.server.types import GhAuthenticationError, GHUnknownException
+from openhands.services.github.github_service import GitHubService
 from openhands.utils.import_utils import get_impl
 
 app = APIRouter(prefix='/api/github')
@@ -20,8 +21,9 @@ async def get_github_repositories(
     sort: str = 'pushed',
     installation_id: int | None = None,
     github_user_id: str | None = Depends(get_user_id),
+    github_user_token: SecretStr | None = Depends(get_github_token),
 ):
-    client = GithubServiceImpl(github_user_id)
+    client = GithubServiceImpl(user_id=github_user_id, token=github_user_token)
     try:
         repos: list[GitHubRepository] = await client.get_repositories(
             page, per_page, sort, installation_id
@@ -44,8 +46,9 @@ async def get_github_repositories(
 @app.get('/user')
 async def get_github_user(
     github_user_id: str | None = Depends(get_user_id),
+    github_user_token: SecretStr | None = Depends(get_github_token),
 ):
-    client = GithubServiceImpl(github_user_id)
+    client = GithubServiceImpl(user_id=github_user_id, token=github_user_token)
     try:
         user: GitHubUser = await client.get_user()
         return user
@@ -66,8 +69,9 @@ async def get_github_user(
 @app.get('/installations')
 async def get_github_installation_ids(
     github_user_id: str | None = Depends(get_user_id),
+    github_user_token: SecretStr | None = Depends(get_github_token),
 ):
-    client = GithubServiceImpl(github_user_id)
+    client = GithubServiceImpl(user_id=github_user_id, token=github_user_token)
     try:
         installations_ids: list[int] = await client.get_installation_ids()
         return installations_ids
@@ -92,8 +96,9 @@ async def search_github_repositories(
     sort: str = 'stars',
     order: str = 'desc',
     github_user_id: str | None = Depends(get_user_id),
+    github_user_token: SecretStr | None = Depends(get_github_token),
 ):
-    client = GithubServiceImpl(github_user_id)
+    client = GithubServiceImpl(user_id=github_user_id, token=github_user_token)
     try:
         repos: list[GitHubRepository] = await client.search_repositories(
             query, per_page, sort, order
