+++ openhands/server/routes/settings.py
@@ -1,11 +1,12 @@
 from fastapi import APIRouter, Request, status
 from fastapi.responses import JSONResponse
+from pydantic import SecretStr
 
 from openhands.core.logger import openhands_logger as logger
 from openhands.server.auth import get_user_id
-from openhands.server.services.github_service import GitHubService
 from openhands.server.settings import GETSettingsModel, POSTSettingsModel, Settings
 from openhands.server.shared import SettingsStoreImpl, config
+from openhands.services.github.github_service import GitHubService
 
 app = APIRouter(prefix='/api')
 
@@ -50,7 +51,7 @@ async def store_settings(
         try:
             # We check if the token is valid by getting the user
             # If the token is invalid, this will raise an exception
-            github = GitHubService(None)
+            github = GitHubService(user_id=None, token=SecretStr(settings.github_token))
             await github.validate_user(settings.github_token)
 
         except Exception as e:
